+++ src/components/app/data/constants.js
@@ -1,3 +1,5 @@
+import { LICENSE_STATUS } from '../../enterprise-user-subsidy/data/constants';
+
 export const emptyRedeemableLearnerCreditPolicies = {
   redeemablePolicies: [],
   learnerContentAssignments: {
@@ -69,3 +71,23 @@ export const ASSIGNMENT_TYPES = {
 export const START_DATE_DEFAULT_TO_TODAY_THRESHOLD_DAYS = 14;
 
 export const SESSION_STORAGE_KEY_LICENSE_ACTIVATION_MESSAGE = 'shouldShowLicenseActivationSuccessMessage';
+
+export const getBaseSubscriptionsData = () => {
+  const baseLicensesByStatus = {
+    [LICENSE_STATUS.ACTIVATED]: [],
+    [LICENSE_STATUS.ASSIGNED]: [],
+    [LICENSE_STATUS.REVOKED]: [],
+  };
+  const baseSubscriptionsData = {
+    subscriptionLicenses: [],
+    customerAgreement: null,
+    subscriptionLicense: null,
+    subscriptionPlan: null,
+    licensesByStatus: structuredClone(baseLicensesByStatus),
+    showExpirationNotifications: false,
+  };
+  return {
+    baseLicensesByStatus,
+    baseSubscriptionsData,
+  };
+};
